feat(pattern_assembler): Breaking up into smaller files
